Added helper functions for creating requests
Also, improved headers

+++ qmi_ctl.c
@@ -1,7 +1,7 @@
 #include "qmi_ctl.h"
 #include "qmi_hdrs.h"
 
-ssize_t ctl_update_cid(int32_t qmi_fd, uint8_t service, uint16_t transaction_id,
+ssize_t ctl_update_cid(int32_t qmi_fd, uint8_t service, uint8_t transaction_id,
         uint8_t release){
 
 }
