+++ qmi_ctl.h
@@ -7,6 +7,6 @@
 //Method for either releasing or updating a CID
 //TODO: Look into transaction ID paramter? Is it needed?
 ssize_t ctl_update_cid(int32_t qmi_fd, uint8_t service,
-        uint16_t transaction_id, uint8_t release);
+        uint8_t transaction_id, uint8_t release);
 
 #endif
